Extract `report` tests
