+++ zrml/prediction-markets/src/old_tests.rs
@@ -138,227 +138,6 @@ fn simple_create_scalar_market(
 }
 
 #[test]
-fn it_allows_to_report_the_outcome_of_a_market() {
-    ExtBuilder::default().build().execute_with(|| {
-        let end = 100;
-        simple_create_categorical_market(
-            Asset::Ztg,
-            MarketCreation::Permissionless,
-            0..end,
-            ScoringRule::Lmsr,
-        );
-
-        let market = MarketCommons::market(&0).unwrap();
-        let grace_period = end + market.deadlines.grace_period;
-        run_to_block(grace_period + 1);
-
-        let market = MarketCommons::market(&0).unwrap();
-        assert_eq!(market.status, MarketStatus::Closed);
-        assert!(market.report.is_none());
-
-        assert_ok!(PredictionMarkets::report(
-            RuntimeOrigin::signed(BOB),
-            0,
-            OutcomeReport::Categorical(1)
-        ));
-
-        let market_after = MarketCommons::market(&0).unwrap();
-        let report = market_after.report.unwrap();
-        assert_eq!(market_after.status, MarketStatus::Reported);
-        assert_eq!(report.outcome, OutcomeReport::Categorical(1));
-        assert_eq!(report.by, market_after.oracle);
-
-        // Reset and report again as approval origin
-        let _ = MarketCommons::mutate_market(&0, |market| {
-            market.status = MarketStatus::Closed;
-            market.report = None;
-            Ok(())
-        });
-
-        assert_ok!(PredictionMarkets::report(
-            RuntimeOrigin::signed(SUDO),
-            0,
-            OutcomeReport::Categorical(1)
-        ));
-    });
-}
-
-#[test]
-fn report_fails_before_grace_period_is_over() {
-    ExtBuilder::default().build().execute_with(|| {
-        let end = 100;
-        simple_create_categorical_market(
-            Asset::Ztg,
-            MarketCreation::Permissionless,
-            0..end,
-            ScoringRule::Lmsr,
-        );
-
-        run_to_block(end);
-
-        let market = MarketCommons::market(&0).unwrap();
-        assert_eq!(market.status, MarketStatus::Closed);
-        assert!(market.report.is_none());
-
-        assert_noop!(
-            PredictionMarkets::report(RuntimeOrigin::signed(BOB), 0, OutcomeReport::Categorical(1)),
-            Error::<Runtime>::NotAllowedToReportYet
-        );
-    });
-}
-
-#[test]
-fn it_allows_only_oracle_to_report_the_outcome_of_a_market_during_oracle_duration_blocks() {
-    ExtBuilder::default().build().execute_with(|| {
-        let end = 100;
-        simple_create_categorical_market(
-            Asset::Ztg,
-            MarketCreation::Permissionless,
-            0..end,
-            ScoringRule::Lmsr,
-        );
-
-        let market = MarketCommons::market(&0).unwrap();
-        let grace_period = end + market.deadlines.grace_period;
-        run_to_block(grace_period);
-
-        let market = MarketCommons::market(&0).unwrap();
-        assert_eq!(market.status, MarketStatus::Closed);
-        assert!(market.report.is_none());
-
-        assert_noop!(
-            PredictionMarkets::report(
-                RuntimeOrigin::signed(CHARLIE),
-                0,
-                OutcomeReport::Categorical(1)
-            ),
-            Error::<Runtime>::ReporterNotOracle
-        );
-
-        assert_ok!(PredictionMarkets::report(
-            RuntimeOrigin::signed(BOB),
-            0,
-            OutcomeReport::Categorical(1)
-        ));
-
-        let market_after = MarketCommons::market(&0).unwrap();
-        let report = market_after.report.unwrap();
-        assert_eq!(market_after.status, MarketStatus::Reported);
-        assert_eq!(report.outcome, OutcomeReport::Categorical(1));
-        assert_eq!(report.by, market_after.oracle);
-    });
-}
-
-#[test]
-fn it_allows_only_oracle_to_report_the_outcome_of_a_market_during_oracle_duration_moment() {
-    ExtBuilder::default().build().execute_with(|| {
-        assert_ok!(PredictionMarkets::create_market(
-            RuntimeOrigin::signed(ALICE),
-            Asset::Ztg,
-            Perbill::zero(),
-            BOB,
-            MarketPeriod::Timestamp(0..100_000_000),
-            get_deadlines(),
-            gen_metadata(2),
-            MarketCreation::Permissionless,
-            MarketType::Categorical(2),
-            Some(MarketDisputeMechanism::SimpleDisputes),
-            ScoringRule::Lmsr
-        ));
-
-        assert_ok!(PredictionMarkets::buy_complete_set(RuntimeOrigin::signed(BOB), 0, CENT));
-
-        // set the timestamp
-        let market = MarketCommons::market(&0).unwrap();
-        // set the timestamp
-
-        set_timestamp_for_on_initialize(100_000_000);
-        run_to_block(2); // Trigger `on_initialize`; must be at least block #2.
-        let grace_period: u64 = market.deadlines.grace_period * MILLISECS_PER_BLOCK as u64;
-        Timestamp::set_timestamp(100_000_000 + grace_period);
-
-        assert_noop!(
-            PredictionMarkets::report(RuntimeOrigin::signed(EVE), 0, OutcomeReport::Categorical(1)),
-            Error::<Runtime>::ReporterNotOracle
-        );
-        assert_ok!(PredictionMarkets::report(
-            RuntimeOrigin::signed(BOB),
-            0,
-            OutcomeReport::Categorical(1)
-        ));
-    });
-}
-
-#[test]
-fn report_fails_on_mismatched_outcome_for_categorical_market() {
-    ExtBuilder::default().build().execute_with(|| {
-        let end = 100;
-        simple_create_categorical_market(
-            Asset::Ztg,
-            MarketCreation::Permissionless,
-            0..end,
-            ScoringRule::Lmsr,
-        );
-        let market = MarketCommons::market(&0).unwrap();
-        let grace_period = end + market.deadlines.grace_period;
-        run_to_block(grace_period + 1);
-        assert_noop!(
-            PredictionMarkets::report(RuntimeOrigin::signed(BOB), 0, OutcomeReport::Scalar(123)),
-            Error::<Runtime>::OutcomeMismatch,
-        );
-        let market = MarketCommons::market(&0).unwrap();
-        assert_eq!(market.status, MarketStatus::Closed);
-        assert!(market.report.is_none());
-    });
-}
-
-#[test]
-fn report_fails_on_out_of_range_outcome_for_categorical_market() {
-    ExtBuilder::default().build().execute_with(|| {
-        let end = 100;
-        simple_create_categorical_market(
-            Asset::Ztg,
-            MarketCreation::Permissionless,
-            0..end,
-            ScoringRule::Lmsr,
-        );
-        let market = MarketCommons::market(&0).unwrap();
-        let grace_period = end + market.deadlines.grace_period;
-        run_to_block(grace_period + 1);
-        assert_noop!(
-            PredictionMarkets::report(RuntimeOrigin::signed(BOB), 0, OutcomeReport::Categorical(2)),
-            Error::<Runtime>::OutcomeMismatch,
-        );
-        let market = MarketCommons::market(&0).unwrap();
-        assert_eq!(market.status, MarketStatus::Closed);
-        assert!(market.report.is_none());
-    });
-}
-
-#[test]
-fn report_fails_on_mismatched_outcome_for_scalar_market() {
-    ExtBuilder::default().build().execute_with(|| {
-        let end = 100;
-        simple_create_scalar_market(
-            Asset::Ztg,
-            MarketCreation::Permissionless,
-            0..end,
-            ScoringRule::Lmsr,
-        );
-        let market = MarketCommons::market(&0).unwrap();
-        let grace_period = end + market.deadlines.grace_period;
-        run_to_block(grace_period + 1);
-        assert_noop!(
-            PredictionMarkets::report(RuntimeOrigin::signed(BOB), 0, OutcomeReport::Categorical(0)),
-            Error::<Runtime>::OutcomeMismatch,
-        );
-        let market = MarketCommons::market(&0).unwrap();
-        assert_eq!(market.status, MarketStatus::Closed);
-        assert!(market.report.is_none());
-    });
-}
-
-#[test]
 fn it_allows_to_dispute_the_outcome_of_a_market() {
     ExtBuilder::default().build().execute_with(|| {
         let end = 2;
@@ -1473,43 +1252,6 @@ fn dispute_emits_event() {
         System::assert_last_event(
             Event::MarketDisputed(0u32.into(), MarketStatus::Disputed, CHARLIE).into(),
         );
-    });
-}
-
-#[test]
-fn it_allows_anyone_to_report_an_unreported_market() {
-    ExtBuilder::default().build().execute_with(|| {
-        let end = 2;
-        simple_create_categorical_market(
-            Asset::Ztg,
-            MarketCreation::Permissionless,
-            0..end,
-            ScoringRule::Lmsr,
-        );
-
-        let market = MarketCommons::market(&0).unwrap();
-        // Just skip to waaaay overdue.
-        run_to_block(end + market.deadlines.grace_period + market.deadlines.oracle_duration + 1);
-
-        assert_ok!(PredictionMarkets::report(
-            RuntimeOrigin::signed(ALICE), // alice reports her own market now
-            0,
-            OutcomeReport::Categorical(1),
-        ));
-
-        let market = MarketCommons::market(&0).unwrap();
-        assert_eq!(market.status, MarketStatus::Reported);
-        assert_eq!(market.report.unwrap().by, ALICE);
-        // but oracle was bob
-        assert_eq!(market.oracle, BOB);
-
-        // make sure it still resolves
-        run_to_block(
-            frame_system::Pallet::<Runtime>::block_number() + market.deadlines.dispute_duration,
-        );
-
-        let market_after = MarketCommons::market(&0).unwrap();
-        assert_eq!(market_after.status, MarketStatus::Resolved);
     });
 }
 
@@ -3238,135 +2980,6 @@ fn on_resolution_correctly_reserves_and_unreserves_bonds_for_advised_approved_ma
     #[cfg(feature = "parachain")]
     ExtBuilder::default().build().execute_with(|| {
         test(Asset::ForeignAsset(100));
-    });
-}
-
-#[test]
-fn report_fails_on_market_state_proposed() {
-    ExtBuilder::default().build().execute_with(|| {
-        assert_ok!(PredictionMarkets::create_market(
-            RuntimeOrigin::signed(ALICE),
-            Asset::Ztg,
-            Perbill::zero(),
-            BOB,
-            MarketPeriod::Timestamp(0..100_000_000),
-            get_deadlines(),
-            gen_metadata(2),
-            MarketCreation::Advised,
-            MarketType::Categorical(2),
-            Some(MarketDisputeMechanism::SimpleDisputes),
-            ScoringRule::Lmsr
-        ));
-        assert_noop!(
-            PredictionMarkets::report(RuntimeOrigin::signed(BOB), 0, OutcomeReport::Categorical(1)),
-            Error::<Runtime>::MarketIsNotClosed,
-        );
-    });
-}
-
-#[test]
-fn report_fails_on_market_state_closed_for_advised_market() {
-    ExtBuilder::default().build().execute_with(|| {
-        assert_ok!(PredictionMarkets::create_market(
-            RuntimeOrigin::signed(ALICE),
-            Asset::Ztg,
-            Perbill::zero(),
-            BOB,
-            MarketPeriod::Timestamp(0..100_000_000),
-            get_deadlines(),
-            gen_metadata(2),
-            MarketCreation::Advised,
-            MarketType::Categorical(2),
-            Some(MarketDisputeMechanism::SimpleDisputes),
-            ScoringRule::Lmsr
-        ));
-        assert_noop!(
-            PredictionMarkets::report(RuntimeOrigin::signed(BOB), 0, OutcomeReport::Categorical(1)),
-            Error::<Runtime>::MarketIsNotClosed,
-        );
-    });
-}
-
-#[test]
-fn report_fails_on_market_state_active() {
-    ExtBuilder::default().build().execute_with(|| {
-        assert_ok!(PredictionMarkets::create_market(
-            RuntimeOrigin::signed(ALICE),
-            Asset::Ztg,
-            Perbill::zero(),
-            BOB,
-            MarketPeriod::Timestamp(0..100_000_000),
-            get_deadlines(),
-            gen_metadata(2),
-            MarketCreation::Permissionless,
-            MarketType::Categorical(2),
-            Some(MarketDisputeMechanism::SimpleDisputes),
-            ScoringRule::Lmsr
-        ));
-        assert_noop!(
-            PredictionMarkets::report(RuntimeOrigin::signed(BOB), 0, OutcomeReport::Categorical(1)),
-            Error::<Runtime>::MarketIsNotClosed,
-        );
-    });
-}
-
-#[test]
-fn report_fails_on_market_state_resolved() {
-    ExtBuilder::default().build().execute_with(|| {
-        assert_ok!(PredictionMarkets::create_market(
-            RuntimeOrigin::signed(ALICE),
-            Asset::Ztg,
-            Perbill::zero(),
-            BOB,
-            MarketPeriod::Timestamp(0..100_000_000),
-            get_deadlines(),
-            gen_metadata(2),
-            MarketCreation::Advised,
-            MarketType::Categorical(2),
-            Some(MarketDisputeMechanism::SimpleDisputes),
-            ScoringRule::Lmsr
-        ));
-        let _ = MarketCommons::mutate_market(&0, |market| {
-            market.status = MarketStatus::Resolved;
-            Ok(())
-        });
-        assert_noop!(
-            PredictionMarkets::report(RuntimeOrigin::signed(BOB), 0, OutcomeReport::Categorical(1)),
-            Error::<Runtime>::MarketIsNotClosed,
-        );
-    });
-}
-
-#[test]
-fn report_fails_if_reporter_is_not_the_oracle() {
-    ExtBuilder::default().build().execute_with(|| {
-        assert_ok!(PredictionMarkets::create_market(
-            RuntimeOrigin::signed(ALICE),
-            Asset::Ztg,
-            Perbill::zero(),
-            BOB,
-            MarketPeriod::Timestamp(0..100_000_000),
-            get_deadlines(),
-            gen_metadata(2),
-            MarketCreation::Permissionless,
-            MarketType::Categorical(2),
-            Some(MarketDisputeMechanism::SimpleDisputes),
-            ScoringRule::Lmsr
-        ));
-        let market = MarketCommons::market(&0).unwrap();
-        set_timestamp_for_on_initialize(100_000_000);
-        // Trigger hooks which close the market.
-        run_to_block(2);
-        let grace_period: u64 = market.deadlines.grace_period * MILLISECS_PER_BLOCK as u64;
-        set_timestamp_for_on_initialize(100_000_000 + grace_period + MILLISECS_PER_BLOCK as u64);
-        assert_noop!(
-            PredictionMarkets::report(
-                RuntimeOrigin::signed(CHARLIE),
-                0,
-                OutcomeReport::Categorical(1)
-            ),
-            Error::<Runtime>::ReporterNotOracle,
-        );
     });
 }
