+++ zrml/prediction-markets/src/tests/mod.rs
@@ -28,6 +28,7 @@ mod manually_close_market;
 mod on_initialize;
 mod on_market_close;
 mod reject_market;
+mod report;
 mod sell_complete_set;
 
 use crate::{mock::*, AccountIdOf, AssetOf, BalanceOf, Config, Error, Event};
